+++ GraphShortestPath.h
@@ -10,9 +10,9 @@ class GraphShortestPath
         GraphShortestPath(const Graph& graph, const unsigned int& node_fro, const unsigned int& node_to);
         virtual ~GraphShortestPath();
         bool is_connected(); // return true if graph is connect and false if not
-        unsigned int dijkstra(); // find the least cost path between node_fro and node_to
+        double dijkstra(); // find the least cost path between node_fro and node_to
         void path_print() const; // print the path found, if any
-        int path_size() const; // return the least cost path distance
+        double path_size() const; // return the least cost path distance
         void path_size_print() const; // print the path distance, if found
 
     protected:
@@ -21,7 +21,7 @@ class GraphShortestPath
         Graph graph;
         unsigned int node_fro;
         unsigned int node_to;
-        int min_distance;
+        double min_distance;
         GraphSortedNodeDistanceList open_set;
         GraphSortedNodeDistanceList closed_set;
 };
